Configure Travis CI, fastlane and SwiftLint
