+++ .bundle/config
@@ -1,0 +1,4 @@
+---
+BUNDLE_RETRY: "3"
+BUNDLE_PATH: "vendor/bundle"
+BUNDLE_JOBS: "4"
