+++ Podfile
@@ -5,7 +5,7 @@ def network_pods
 end
 
 def dev_pods
-  pod 'SwiftLint', '~> 0.37'
+  pod 'SwiftLint', '~> 0.35'
 end
 
 def test_pods
@@ -17,15 +17,13 @@ def util_pods
     pod 'Kingfisher', '~> 5.10'
 end
 
-target 'Surveys-CodeChallenge' do
+target 'SurveysCodeChallenge' do
   use_frameworks!
   network_pods
   util_pods
   dev_pods
 
-  # Pods for Surveys-CodeChallenge
-
-  target 'Surveys-CodeChallengeUITests' do
+  target 'SurveysCodeChallengeTests' do
     inherit! :search_paths
     dev_pods
     test_pods
