+++ SurveysCodeChallenge.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Surveys-CodeChallenge.xcodeproj">
+      location = "self:SurveysCodeChallenge.xcodeproj">
    </FileRef>
 </Workspace>
