+++ SurveysCodeChallenge.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:Surveys-CodeChallenge.xcodeproj">
+      location = "group:SurveysCodeChallenge.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
