+++ SurveysCodeChallenge/AppDelegate.swift
@@ -1,0 +1,20 @@
+//
+//  AppDelegate.swift
+//  SurveysCodeChallenge
+//
+//  Created by Su Nguyen on 11/24/19.
+//  Copyright © 2019 Su Nguyen. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
