+++ SurveysCodeChallenge/Controllers/SurveyDetail/SurveyDetailViewModel.swift
@@ -9,5 +9,4 @@
 import Foundation
 
 final class SurveyDetailViewModel {
-    
 }
