+++ SurveysCodeChallenge/Controllers/Surveys/SurveysViewController.swift
@@ -12,9 +12,5 @@ final class SurveysViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
-
