+++ SurveysCodeChallenge/Resources/Assets.xcassets/Contents.json
