+++ SurveysCodeChallenge/Storyboards/Base.lproj/Main.storyboard
@@ -13,12 +13,12 @@
         <!--Surveys View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="SurveysViewController" customModule="Surveys_CodeChallenge" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="SurveysViewController" customModule="SurveysCodeChallenge" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="SurveysViewController" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ru0-HT-apK">
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="SurveysViewController" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="xPa-Zw-HtP">
                                 <rect key="frame" x="121" y="437.5" width="172.5" height="21"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <nil key="textColor"/>
@@ -27,8 +27,8 @@
                         </subviews>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                         <constraints>
-                            <constraint firstItem="ru0-HT-apK" firstAttribute="centerX" secondItem="8bC-Xf-vdC" secondAttribute="centerX" id="P8L-wi-fWV"/>
-                            <constraint firstItem="ru0-HT-apK" firstAttribute="centerY" secondItem="8bC-Xf-vdC" secondAttribute="centerY" id="wvQ-xO-Z50"/>
+                            <constraint firstItem="xPa-Zw-HtP" firstAttribute="centerX" secondItem="8bC-Xf-vdC" secondAttribute="centerX" id="Qax-jQ-8Ue"/>
+                            <constraint firstItem="xPa-Zw-HtP" firstAttribute="centerY" secondItem="8bC-Xf-vdC" secondAttribute="centerY" id="R5b-hQ-xaS"/>
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
