+++ fastlane/Scanfile
@@ -1,0 +1,4 @@
+scheme("SurveysCodeChallengeTests")
+device("iPhone 8")
+clean(true)
+skip_build(true)
