+++ scripts/install
@@ -1,0 +1,36 @@
+#!/bin/bash
+
+# DEV ENV
+
+# homebrew
+if ! which brew > /dev/null; then
+    sudo chown -R "$(whoami)":admin '/usr/local'
+    /usr/bin/ruby -e "$(curl -fsSL 'https://raw.githubusercontent.com/Homebrew/install/master/install')"
+    mkdir -p '/Library/Caches/Homebrew'
+    sudo chown -R "$(whoami)":admin '/Library/Caches/Homebrew'
+fi
+
+brew update
+
+# rbenv if need
+if [[ "$CI" != 'true' ]]; then
+    if ! which rbenv > /dev/null; then
+        echo 'install rbenv'
+        brew install rbenv
+        eval "$(rbenv init -)"
+        echo 'which rbenv > /dev/null && eval "$(rbenv init -)"' >> ~/.bashrc
+    fi
+    brew outdated ruby-build || brew upgrade ruby-build
+    echo 'n' | rbenv install || true
+    rbenv rehash
+fi
+
+gem install bundler
+
+# PROJECT DEPENDENCES
+
+bundle install --path=vendor/bundle --jobs 4 --retry 3
+
+if ! bundle exec pod install; then
+    bundle exec pod install --repo-update
+fi
