+++ scripts/lint
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+echo "TRAVIS_REPO_SLUG = $TRAVIS_REPO_SLUG"
+echo "TRAVIS_PULL_REQUEST = $TRAVIS_PULL_REQUEST"
+
+if [ "$TRAVIS_PULL_REQUEST" != "false" ]
+then
+  ./Pods/SwiftLint/swiftlint --reporter json > swiftlint-report.json || false
+  bundle exec linterbot $TRAVIS_REPO_SLUG $TRAVIS_PULL_REQUEST < swiftlint-report.json
+fi
